añadidas las queries que faltaban

+++ Acme-Chorbies-2.0/src/main/java/repositories/TasteRepository.java
@@ -13,9 +13,9 @@ import domain.Taste;
 public interface TasteRepository extends JpaRepository<Taste, Integer> {
 
 	@Query("select (select sum(t.stars) from Taste t where t.chorbi.id=c.id) as auxiliar from Chorbi c order by auxiliar ASC")
-	Collection<Integer> minMaxStars();
+	Collection<Long> minMaxStars();
 
 	@Query("select sum(t.stars)*1.0/(select count(c)*1.0 from Chorbi c) from Taste t")
-	Long avgStars();
+	Double avgStars();
 
 }
